requirements.txt
Django>=3.2,<3.3.0
djangorestframework>=3.12.2,<3.13.0
pandas
tensorflow